+++ src/rules/index.ts
@@ -18,6 +18,7 @@ import { invalidEnvVarNameRule } from "./invalid-env-var-name.js";
 import { emptyEnvValueRule } from "./empty-env-value.js";
 import { cwdNotAbsoluteRule } from "./cwd-not-absolute.js";
 import { secretInArgsRule } from "./secret-in-args.js";
+import { duplicateImageRule } from "./duplicate-image.js";
 
 export const BUILTIN_RULES: Rule[] = [
   structureRules,
@@ -39,4 +40,5 @@ export const BUILTIN_RULES: Rule[] = [
   emptyEnvValueRule,
   cwdNotAbsoluteRule,
   secretInArgsRule,
+  duplicateImageRule,
 ];